Exercise model and form completion
Re-created exercise model using ints instead of floats. Removed type of
weight from exercise model. Finished form view. Need to hook up ajax
submission.

+++ db/migrate/20120321204905_create_exercises.rb
@@ -2,9 +2,9 @@ class CreateExercises < ActiveRecord::Migration
   def change
     create_table :exercises do |t|
       t.string :name
-      t.float :weight_each_side
-      t.float :bar_weight
-      t.float :total_weight
+      t.integer :weight_each_side
+      t.integer :bar_weight
+      t.integer :total_weight
       t.string :weight_units
       t.string :type_of_weight
       t.string :reps
